cdk/requirements.txt
# Add CDK dependencies here
aws-cdk-lib
constructs>=10.0.0
